+++ src/components/product-manage/product-manage-tab.ts
@@ -1,7 +1,7 @@
-import Component from '../abstract/component';
-import { customElement } from '../decorators/decortators';
-import Store from '../flux/store';
-import { Tab } from '../types';
+import Component from '../../abstract/component';
+import { customElement } from '../../decorators/decortators';
+import Store from '../../flux/store';
+import { Tab } from '../../types';
 import './add-product-form';
 import './product-inventory';
