+++ src/components/purchase-product/product-menu.ts
@@ -1,10 +1,10 @@
-import Component from '../abstract/component';
-import { ACTION } from '../constants';
-import { customElement } from '../decorators/decortators';
-import createAction from '../flux/createAction';
-import Store from '../flux/store';
-import { EventOnElement, ProductItem } from '../types';
-import { validatePurchaseProduct } from '../validation/validators';
+import Component from '../../abstract/component';
+import { ACTION } from '../../constants';
+import { customElement } from '../../decorators/decortators';
+import createAction from '../../flux/createAction';
+import Store from '../../flux/store';
+import { EventOnElement, ProductItem } from '../../types';
+import { validatePurchaseProduct } from '../../validation/validators';
 
 @customElement('product-menu')
 class ProductMenu extends Component {
