+++ src/scss/components/product-manage/_product-inventory.scss
