+++ src/scss/components/purchase-product/_insert-money-form.scss
